+++ packages/webpack-config-terra/tests/jest/webpack.config.test.js
@@ -216,11 +216,13 @@ describe('webpack config', () => {
     //  and adds to dev server options', () => {
     const expectedOuput = {
       hot: false,
-      inline: false,
       host: '0.0.0.0',
-      stats: {
-        colors: true,
-        children: false,
+      devMiddleware: {
+        publicPath: '',
+        stats: {
+          children: false,
+          colors: true,
+        },
       },
     };
     expect(config.devServer).toEqual(expect.objectContaining(expectedOuput));