+++ packages/terra-functional-testing/CHANGELOG.md
@@ -1,8 +1,11 @@
 # Changelog
 
 ## Unreleased
+* Added
+  * Added devMiddleWare to support webpack-dev-server v4
 
 * Added
+  * Added authentication for accessing screenshots from the remote site.
   * Added `useRemoteReferenceScreenshots` cli option for downloading reference screenshots from a remote site for screenshot comparisons.
 
 ## 2.7.0 - (February 11, 2022)