+++ packages/terra-toolkit-docs/CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## Unreleased
+* Added
+  * Added upgrade guide for webpack-config-terra
 
 ## 2.9.0 - (February 11, 2022)